backend/test_gemini.py
"""
Quick test script for the new Gemini-powered Talk2PDF backend
Run this after setting GEMINI_API_KEY in .env file
"""

import os
import sys
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

def test_lightweight_system():
    """Test the lightweight ingestion and Gemini query system"""
    
    print("🧪 Testing Talk2PDF Gemini Integration")
    print("=" * 50)
    
    # Check API key
    if not os.getenv("GEMINI_API_KEY"):
        print("❌ GEMINI_API_KEY not found in environment variables")
        print("💡 Create a .env file in the backend directory with:")
        print("   GEMINI_API_KEY=your_key_here")
        return False
    
    try:
        # Test imports
        print("📦 Testing imports...")
        from lightweight_ingest import lightweight_ingest_document
        from gemini_query import answer_query_gemini
        print("✅ All imports successful")
        
        # Test sample document processing (if exists)
        sample_pdf = "docs/sample.pdf"
        if os.path.exists(sample_pdf):
            print(f"📄 Testing with {sample_pdf}")
            
            doc_id = "test_doc_123"
            success = lightweight_ingest_document(sample_pdf, doc_id)
            
            if success:
                print("✅ Document processing successful")
                
                # Test query
                print("🤖 Testing Gemini query...")
                result = answer_query_gemini(doc_id, "What is this document about?")
                
                print(f"📋 Answer: {result['answer'][:100]}...")
                print(f"🎯 Confidence: {result['confidence']}")
                print(f"📚 Sources: {len(result['sources'])} found")
                print("✅ Gemini query successful")
                
                return True
            else:
                print("❌ Document processing failed")
                return False
        else:
            print(f"⚠️  Sample PDF not found at {sample_pdf}")
            print("✅ System imports working - ready for real documents")
            return True
            
    except Exception as e:
        print(f"❌ Error: {str(e)}")
        return False

def test_api_key_only():
    """Just test if Gemini API key works"""
    try:
        import google.generativeai as genai
        
        api_key = os.getenv("GEMINI_API_KEY")
        if not api_key:
            print("❌ No API key found")
            return False
            
        genai.configure(api_key=api_key)
        model = genai.GenerativeModel('gemini-1.5-flash')
        
        response = model.generate_content("Say 'API working' if you can read this.")
        print(f"🤖 Gemini response: {response.text}")
        print("✅ Gemini API key is working!")
        return True
        
    except Exception as e:
        print(f"❌ Gemini API test failed: {e}")
        return False

if __name__ == "__main__":
    print("🚀 Talk2PDF Gemini Test Suite")
    print("\n1. Testing API Key...")
    api_works = test_api_key_only()
    
    if api_works:
        print("\n2. Testing Full System...")
        system_works = test_lightweight_system()
        
        if system_works:
            print("\n🎉 All tests passed! Your system is ready for deployment.")
        else:
            print("\n⚠️  Some tests failed, but API key works.")
    else:
        print("\n❌ API key test failed. Please check your GEMINI_API_KEY.")
    
    print("\n📊 Estimated deployment size: ~50MB (vs 500MB+ with transformers)")
    print("💰 Cost: Free tier should handle moderate usage")
    print("⚡ Speed: Much faster startup and processing")
