backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
google-generativeai>=0.3.0
pdfplumber>=0.10.3
pillow>=10.1.0
numpy>=1.24.3
scikit-learn>=1.3.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
